+++ fraud-service/src/main/java/com/muazwzxv/fraud/FraudCheckRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 import java.util.UUID;
 
-public interface FraudCheckRepository extends JpaRepository<FraudCheckHistory, Integer> {
+public interface FraudCheckRepository extends JpaRepository<FraudCheckHistory, Long> {
     Optional<FraudCheckHistory> findByUuid(UUID uuid);
 }
