+++ fraud-service/src/main/java/com/muazwzxv/fraud/FraudCheckService.java
@@ -1,6 +1,6 @@
 package com.muazwzxv.fraud;
+
 import lombok.AllArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
@@ -11,7 +11,7 @@ public class FraudCheckService {
 
     private final FraudCheckRepository fraudCheckRepository;
 
-    public boolean isFraudulentCustomer(Integer customerId) {
+    public boolean isFraudulentCustomer(Long customerId) {
         // We'll say everyone is not a fraud right now
         fraudCheckRepository.save(
                 FraudCheckHistory.builder()
